Move Linear Operators
