+++ mri/operators/linear/dictionary.py
