+++ mri/operators/linear/utils.py
