+++ mri/operators/linear/wavelet.py
